Added use of event-navigation library and renamed bloc and repository files.
