--- lib/bloc/navigation/navigation.dart
@@ -1,12 +1,0 @@
-import 'package:event_navigation/event_navigation.dart';
-
-const POSSIBLE_NAVIGATION = {'home', 'play', 'workout', 'exercise', 'settings'};
-
-MainNavigationBloc<String> generateNavigationBloc() {
-  return MainNavigationBloc(
-    defaultNavigation: 'home',
-    strategy: ListNavigationStrategy(
-        possibleNavigations: POSSIBLE_NAVIGATION.toList()),
-    undoStrategy: UndoRedoMainNavigationStrategy(),
-  );
-}
