+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:event_bloc/event_bloc.dart';
 import 'package:event_navigation/event_navigation.dart';
 import 'package:flutter/material.dart';
-import 'package:simple_exercise_builder/bloc/navigation/navigation.dart';
+import 'package:simple_exercise_builder/bloc/navigation.dart';
 import 'package:simple_exercise_builder/bloc_layer.dart';
 import 'package:simple_exercise_builder/repository_layer.dart';
 import 'package:simple_exercise_builder/screen/main.dart';
@@ -22,11 +22,11 @@ class MyApp extends StatelessWidget {
 class AppLayer extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return EventNavigationApp(
       title: 'Exercise Demo',
       theme: ThemeData(highlightColor: Colors.lightBlue[100]),
       darkTheme: ThemeData.dark(),
-      home: const MyHomePage(title: 'Exercise Demo Home Page'),
+      child: const MyHomePage(title: 'Exercise Demo Home Page'),
     );
   }
 }
