+++ lib/repository_layer.dart
@@ -1,6 +1,6 @@
 import 'package:event_bloc/event_bloc.dart';
 import 'package:flutter/widgets.dart';
-import 'package:simple_exercise_builder/repository/exercise_repo.dart';
+import 'package:simple_exercise_builder/repository/exercise/exercise_repo.dart';
 
 class RepositoryLayer extends StatelessWidget {
   final Widget child;
