+++ pubspec.yaml
@@ -15,8 +15,7 @@ dependencies:
   carousel_slider: ^4.0.0
   cupertino_icons: ^1.0.2
   event_bloc: ^1.0.0+1
-  event_navigation: 
-    path: ../event-navigation
+  event_navigation: ^0.1.0
   hive: ^2.0.4
   provider: ^5.0.0
   search_me_up: ^0.1.0
